+++ app_fempic_cg/omp/deposit_charge_on_nodes_kernel_loop.hpp
@@ -57,8 +57,8 @@ void opp_par_loop_all__deposit_charge_on_nodes_kernel(opp_set set, opp_iterate_t
     #pragma omp parallel for 
     for (int thr = 0; thr < nthreads; thr++)
     {
-        const size_t start  = (iter_size * thr) / nthreads;
-        const size_t finish = (iter_size * (thr+1)) / nthreads;
+        const size_t start  = ((size_t)iter_size * thr) / nthreads;
+        const size_t finish = ((size_t)iter_size * (thr+1)) / nthreads;
 
         OPP_REAL* arg1_dat_thread_data = (OPP_REAL *)((*(args[1].dat->thread_data))[thr]);
