+++ app_fempic_cg/omp/inject_ions_kernel_loop.hpp
@@ -19,7 +19,7 @@ inline void inject_ions_kernel(
 {
     double a = dummy_part_random[0];
     double b = dummy_part_random[1];
-    if ((a + b) > 1)
+    if ((a + b) > 1)  // TODO : Change the random dat to avoid this
     {
         a = (1 - a);
         b = (1 - b);
@@ -84,8 +84,8 @@ void opp_par_loop_injected__inject_ions_kernel(opp_set set, opp_iterate_type,
     #pragma omp parallel for 
     for (int thr = 0; thr < nthreads; thr++)
     {
-        const size_t start  = (iter_size * thr) / nthreads;
-        const size_t finish = (iter_size * (thr+1)) / nthreads;
+        const size_t start  = ((size_t)iter_size * thr) / nthreads;
+        const size_t finish = ((size_t)iter_size * (thr+1)) / nthreads;
       
         for (size_t n = start; n < finish; n++)
         {
